Splitting out tracker code
